# Optimizations in notify-one