--- a/unreleased_history/performance_improvements/parallel_notify-one_i_LaunchParallelMemTableWriters.md
+++ b/unreleased_history/performance_improvements/parallel_notify-one_i_LaunchParallelMemTableWriters.md
@@ -1,0 +1,1 @@
+Improved write throughput to memtable when there's a large number of concurrent writers and allow_concurrent_memtable_write=true(#12545)