@cbi42 2024-05-10
Hi thanks for the PR. The change makes sense to me. I wonder if you have results for the performance improvement for each step (1. Parallelize SetState vs 2. Wake up non-STATE_LOCKED_WAITING first). I did some benchmark with fillrandom and I saw 1. improves performance but 2. does not show much improvement. 

```
Command:
./db_bench --benchmarks=fillrandom[-X5] --threads=160 --seed=1708494134896523 --duration=30 --disable_auto_compactions=1 --max_write_buffer_number=100 --min_write_buffer_number_to_merge=100 --statistics=0

main:
fillrandom [AVG    5 runs] : 130401 (± 535) ops/sec;   14.4 (± 0.1) MB/sec
fillrandom [MEDIAN 5 runs] : 130207 ops/sec;   14.4 MB/sec

this PR: 
fillrandom [AVG    5 runs] : 282247 (± 70009) ops/sec;   31.2 (± 7.7) MB/sec
fillrandom [MEDIAN 5 runs] : 317884 ops/sec;   35.2 MB/sec

this PR without 2.
fillrandom [AVG    5 runs] : 281672 (± 70145) ops/sec;   31.2 (± 7.8) MB/sec
fillrandom [MEDIAN 5 runs] : 317164 ops/sec;   35.1 MB/sec
```