@casualwind 2024-05-13
@cbi42 Thank you for your review!
for each step :1. Parallelize SetState  2. Wake up non-STATE_LOCKED_WAITING first

We have tested the performance before(default --enable_pipelined_write=1), and the step2‘s impact on performance is indeed nearly to none.

We put them together because in our previous tests ( the main branch is not the newest), when --enable_pipelined_write=0, step2+step1 got better performance when there is only step1. Now we tested them on the latest main branch, step2 does have no impact on performance at any time.

Currently, we only need the step 1 rather than step1+step2.