@cbi42 2024-05-14
> We wonder why their gap so much?



> @cbi42 We have a doubt besides.
> 
> there are 2 commands:
> 
> ```
> Command 1:
> ./db_bench --benchmarks=fillrandom[-X5] --threads=160 --seed=1708494134896523 --duration=30 --disable_auto_compactions=1 --max_write_buffer_number=100 --min_write_buffer_number_to_merge=100 --statistics=0
> 
> Command 2:
> ./db_bench --benchmarks=fillrandom --threads=160 --seed=1708494134896523 --duration=30 --disable_auto_compactions=1 --max_write_buffer_number=100 --min_write_buffer_number_to_merge=100 --statistics=0
> ```
> 
> We run command 1: stdev/average is nearly 25% We run command 2 for 5 times: stdev/average is under 1%.
> 
> We wonder why their gap so much?

I just realize that fillrandom[-X5] will reuse the same DB, so it's probably better to just run fillrandom without the [-X5].