@cbi42 2024-05-20
> @cbi42 We have updated the PR summary and added a change log.

Hi - for PR summary, I still see two optimizations listed, it should only have the first one now?